allows cluster to setup port forwarding for testing e.g. kafka (2) print exec result in log (3) renamed packages
